odrs: Add caching support to fetch and ratings endpoints
Particularly with the ratings endpoint, this will allow users to avoid
re-downloading all the ratings data if their local cache is up to date.

Note that in production, this is already handled by the WSGI proxy
layer. It helps to be explicit in the code about what caching is
suitable for each endpoint, though.

Signed-off-by: Philip Withnall <withnall@endlessm.com>