httpd: Support content encoding and pre-compressed forms of ratings.json
`ratings.json` is quite large (currently 827598B). When gzipped, it
drops to 103478B, and when brotli-encoded it drops to 81655B (a factor
of 10 smaller than the original file).

Since this is one of the most commonly-requested ODRS endpoints, it
probably will help our bandwidth bills (and client download times) to
allow compression on it, using the `Content-Encoding` HTTP header.

brotli compression, in particular, is not very fast (takes around 1s to
compress `ratings.json`), so we should cache the pre-compressed files on
the server.

This is based off the following documentation:
 - https://httpd.apache.org/docs/2.4/mod/mod_brotli.html

Signed-off-by: Philip Withnall <withnall@endlessm.com>